Create new tests for the new regions code, rename old test file
